Refactor scheduled processing, fixes #309
Update job processing within timed sets to work on fixed
job counts and remove them one at a time as they are actually
processed. This will minimize any possible job loss due
to errors and ensure that huge job counts don't lead to Redis
timeouts.

+++ manager/manager.go
@@ -92,7 +92,7 @@ type Manager interface {
 
 	WorkingCount() int
 
-	ReapExpiredJobs(when time.Time) (int, error)
+	ReapExpiredJobs(when time.Time) (int64, error)
 
 	// Purge deletes all dead jobs
 	Purge(when time.Time) (int64, error)
